refactor(cli): split Rust-backed export formats out of export.ts, revert budget raise
The reviewer correctly flagged that this PR's fix grew export.ts from 431 to
454 lines, 18 over its recorded module-size-allowlist.txt budget of 436, and
that check-module-size.mjs failed despite the PR body claiming it passed. The
follow-up commit "fixed" this by raising the allowlist budget to 454 instead
-- the exact escape hatch check-module-size.mjs's own message warns against
("Shrink or split instead of raising the budget").

Revert that budget raise, and instead extract the Rust-backed exporters
(obj/gltf/glb/jsonld/ifcx/usd/step, and the shared wasm GeometryProcessor
bootstrap they use) into a new sibling module, export-rust-formats.ts. That
block is the one cohesive seam in the file: every import it needs
(GeometryProcessor, isNoRenderGeometryError, countGlbMeshes, hasFlag,
formatGeometryReport/NO_DIAGNOSTICS_LINE, readFile) is used nowhere else in
export.ts.

export.ts is now 322 lines -- under the 400-line house limit, so its
allowlist row is deleted rather than lowered, per check-module-size.mjs's own
convention. export-rust-formats.ts is 167 lines, also under the limit.
check-module-size.mjs now passes with 339 allowlisted files (down from 340)
and no allowlist increase.

Pure refactor, no behaviour change: pnpm --filter @ifc-lite/cli test is still
784 passed / 8 skipped, and mutating each of #4044's fix conditions back to
its broken form (the unfiltered-export refs pass-through, and the zero-match
--type fatal() guard) reproduces the original defects and fails the relevant
test, confirming both survived the extraction.

Also amends the PR's changeset and corrects its body's Test plan, which
claimed check-module-size.mjs passed when it did not.
