+++ scripts/module-size-allowlist.txt
@@ -194,7 +194,6 @@
    433 packages/clash/src/duplicates.ts
    547 packages/cli/src/commands/ask.ts
    495 packages/cli/src/commands/create.ts
-   454 packages/cli/src/commands/export.ts
    518 packages/cli/src/commands/extract-entities.ts
    609 packages/cli/src/commands/query.ts
    791 packages/cli/src/headless-backend.ts
